refactor: api 함수 도메인 별 분리
